Inline interpretation helpers into test_interpretation_common
