+++ tests/unit/disable_jit/test_ground_rule_helpers.py
@@ -1,6 +1,6 @@
 import pytest
 from unittest.mock import Mock, call
-from tests.unit.disable_jit.interpretation_helpers import get_interpretation_helpers
+from tests.unit.disable_jit.test_interpretation_common import get_interpretation_helpers
 
 # Preload defaults so decorators resolve
 _default = get_interpretation_helpers("interpretation_fp")
